Validate BASE32 secret and support large secrets
All invalid characters in the string were interpreted as zero,
same as the equal ('=') padding.
For short strings treating padding as zero is fine ad the HMAC
algorithm pad the secret with zeroes if it fit into the
block. However if it does not fit the computation get wrong.
If the BASE32 secret is pretty large (104 characters for SHA1),
the key is replaced by an hashed version of it.
So, to make an example if the secret is up to 103 "A" characters
the binary key will be the same but adding any more "A" will cause the
key to change (and so the tokens).
This requires us to pass a key that is not multiple of 5 bytes,
so the implementation of otp_byte_secret() was changed to satisfy
this.
While at it catch and report invalid characters.
OTP_DEFAULT_BASE32_OFFSETS array was moved to C file, no reasons
to expose it.
Padding "=" characters are stripped at the beginning, allowing
to pass them or not, as most implementations are allowing this.

Signed-off-by: Frediano Ziglio <freddy77@gmail.com>